Fix sketch computation bug
Basically, there was an extra "i+1" in the loop that windows the data
for sketches.  This means that every other reading within the
fetch_window was being ignored, which is especially problematic when
you your sampling and resampling rates are the same, since you'll then
be missing every other bucket.